Implement character creation UI and logic for attribute allocation

+++ src/core/races.rs
@@ -12,43 +12,43 @@ pub enum Race {
 }
 
 impl Race {
-    pub fn modifier(&self, attr: Attribute) -> f32 {
+    pub fn modifier(&self, attr: Attribute) -> i8 {
         match attr {
             Attribute::Strength => match self {
-                Race::Dwarf => 1.0,
-                Race::Elf => -2.0,
-                Race::Human => 0.0,
-                Race::Orc => 2.0,
+                Race::Dwarf => 1,
+                Race::Elf => -2,
+                Race::Human => 0,
+                Race::Orc => 2,
             },
             Attribute::Dexterity => match self {
-                Race::Dwarf => -1.0,
-                Race::Elf => 2.0,
-                Race::Human => 1.0,
-                Race::Orc => 0.0,
+                Race::Dwarf => -1,
+                Race::Elf => 2,
+                Race::Human => 1,
+                Race::Orc => 0,
             },
             Attribute::Constitution => match self {
-                Race::Dwarf => 2.0,
-                Race::Elf => -1.0,
-                Race::Human => 0.0,
-                Race::Orc => 2.0,
+                Race::Dwarf => 2,
+                Race::Elf => -1,
+                Race::Human => 0,
+                Race::Orc => 2,
             },
             Attribute::Intelligence => match self {
-                Race::Dwarf => 0.0,
-                Race::Elf => 1.0,
-                Race::Human => 0.0,
-                Race::Orc => -1.0,
+                Race::Dwarf => 0,
+                Race::Elf => 1,
+                Race::Human => 0,
+                Race::Orc => -1,
             },
             Attribute::Wisdom => match self {
-                Race::Dwarf => 1.0,
-                Race::Elf => 1.0,
-                Race::Human => 0.0,
-                Race::Orc => 0.0,
+                Race::Dwarf => 1,
+                Race::Elf => 1,
+                Race::Human => 0,
+                Race::Orc => 0,
             },
             Attribute::Charisma => match self {
-                Race::Dwarf => -1.0,
-                Race::Elf => 1.0,
-                Race::Human => 1.0,
-                Race::Orc => -1.0,
+                Race::Dwarf => -1,
+                Race::Elf => 1,
+                Race::Human => 1,
+                Race::Orc => -1,
             },
         }
     }
